Added executables to start the proxy and server.
Rackup is used to start the servers. It is done "cleanly" meaning the
$LOAD_PATH is not modified and rubygems is not explicitly required.
Well, it kind of is because it's gem indexer is used.

There is one problem with the above. Rackup daemonize does not work. The
app and lib dirs are lost from the $LOAD_PATH so the apps fail to load.
It probably has to do with something I don't understand about fork.